Remove test duplication
Consolidate tests for bad input into a single file instead of having
the same code in two files.
